Fixed OutOfMemory exception hashing large files.
Apparently, in 0.0.10.0, when we are hashing the large file, it will take up almost the entire RAM of the computer, and then the command crashes with OutOfMemory exception. That's because we were taking a string of the content of the files and then hashing them, which has caused the memory leak.

We have spotted this error and removed it in this version so that hashing large files is now faster! No OutOfMemory exception!