Added sending mail, making the IMAP shell the mail shell.
Also, fixed port parsing. Currently, the SMTP connection is port 587, assuming it's encrypted with STARTTLS. This is true for most SMTP servers. If you're logged in to some rare SMTP server that doesn't support any type of encryption, report it to Issues, and we'll try to add another prompt that talks about encryption.