Reworked on settings -> Shell -> Custom Colors to use...
...ColorWheel()

As part of this change, we've changed the return type for this function from Integer to String to support both 255 color mode and true color mode. This signifies that both ThemeStudio and settings will be able to set colors in true color mode.

As part of this process, we've obsoleted SMultivar as it isn't being used anymore.

Even if it seems to work, the user will still have to modify the .ini file so the semicolons aren't treated as comments by enclosing the VT sequence used by double quotes like this: "0;255;255".

We'll make an Extensification dedicated to that. Meanwhile, internal tests will be made.