Added dark color detection
For now, it's only compatible with true color Color class instances. It's going to be changed in the future.

Meanwhile, we've fixed ColTypes.Gray on text writers mistreating light background colors as dark colors, leading to light-to-light color writing. Please note that it only works on true color mode for the reasons above.

More information found internally.