JSON-ified KS localizations
Once KS 0.0.16.0 releases, we'll provide instructions on how to use all the tools in the wiki. For now, such information have to stay internal.

Why the Translations folder exist in the same place that Kernel Simulator exists? Because this folder provides the translations text files used for KSJsonifyLocales, and are not meant to be used by Kernel Simulator because it uses the embedded JSON files.

Define COPYTOKS to copy the generated locates to Kernel Simulator's resources directly. We've disabled it by default just to be safe.