Added new API SetConsoleColor()
This makes use of the new Color class to handle setting the console color, whether it's a standard 16-bit color that can be already set using the Console.{Fore|Back}groundColor property. It's a public function so any mod or custom screensaver can use it.

If running on Windows, it only sets the appropriate console color without having to set the CursorLeft to its correct value. The reason is that because Mono sets the wrong console cursor left position when the VT sequences are encountered. This causes Console.CursorLeft to have the value of the current cursor position plus the VT sequences' string lengths when the terminal emulator still shows the correct cursor position. This doesn't happen in .NET Framework on Windows.

The workaround is to grab the old CursorLeft and CursorTop positions prior to writing the VT sequence used to set the console colors. This will fix the following:

- Separators are incomplete
- Dissolve generates incomplete random color (EB#2)

The other workaround is to subtract the CursorLeft value by the length of the VT sequence string, but it's a bad idea because you'll never know if you're going to deal with the text that's going to be wrapped on the end of the line buffer, especially on calls to any TextWriters. For example, when we tried this workaround on the Dissolve screensaver, after few spaces, we got this exception:

System.ArgumentOutOfRangeException: Value must be positive and below the buffer width. Parameter name: left

So using the OldLeft and OldTop variables, we're safe from these kind of weird errors.

The other workaround we had to use for the SetConsoleColor() function is to replace all instances of vbNewLine with Environment.NewLine using the field on Kernel that impersonates vbNewLine, thus further lessening the usage of Microsoft.VisualBasic namespace. This fixes the extra newlines that are generated from the SetConsoleColor() function thanks to setting the console position in KS builds that are built by Visual Studio on Windows.

The extra newlines that are generated by making separators are still there, but it could be fixed.