Fixed printing being VERY SLOW on slow processors
This is the 1,400th commit we've done to this project!

We've used the CPU profiling feature in VS and saw that ConvertVTSequences took a few CPU cycles to do its job. It's up to Extensification.

We've switched to a more-simple initialization of VTSequenceForeground and VTSequenceBackground properties, and it seemed to reduce the need of CPU cycles.

However, this falls more to ConsoleColorsInfo constructor when we had to make an object reference to a parsed ColorDataJson. If this is run in a loop, it will cause performance drops due to the high CPU cycles.

So, we've made the public read-only JSON token of ConsoleColorsData. Any mod or screensaver can use it.

KS 0.0.16 RC1 before this change (time) on AMD Athlon XP 1500+ (startup, login, shutdown):

real    0m35.750s
user    0m19.908s
sys     0m2.255s

KS 0.0.16 RC1 after this change (time) on AMD Athlon XP 1500+ (startup, login, shutdown):

real    0m21.480s
user    0m6.525s
sys     0m1.805s

NOTICE: If you're running this build of KS and you're comparing it to the previous build of KS on high-end processors such as Intel Core i7-8700, you'll see next-to-zero difference in performance.