EditLine now behaves slightly different for cross-platform
Before this version, we've used the SendWait() routine from the Windows.Forms namespace. However, it didn't work for us in the Linux version and it proved to be slow in Windows systems, so we'll use the brand new ReadLine.Reboot utility with the Default value to specify the default value if the line isn't changed.

However, this warrants the behavioral change to this command in the text editor. Don't worry; if you want the line unchanged, just press the ENTER key, and the line will not be changed.