Cleaned up unused deps, including DirectoryServices
We've used DirectoryServices in the past (0.0.3) to implement the commands which listed all Windows computers or computers that had the SAMBA share installed, like "lsnet" and "lsnettree". This file was used to do such commands:

https://github.com/EoflaOE/Kernel-Simulator/blob/v0.0.3-alpha/Kernel%20Simulator/Network/NetworkList.vb

However, it has proven to be unreliable, so it was removed in 0.0.7 to comply with the compatibility with Linux due to DirectoryServices not working in such systems.