Fixed screensaver not starting on .NET 6.0 in Linux systems
Apparently, Console class as a whole in modern .NET is behaving differently from .NET Framework and is not thread safe. It can be traced back to the first version of .NET Core 1.0 (2016).

More info:
  - https://blog.mavnn.co.uk/system-dot-console-is-why-we-cant-have-nice-things/

Unrelated, but it has to do with Console.WriteLine: https://putridparrot.com/blog/console-readkey-and-console-writeline-deadlock/

We didn't experience the Console.WriteLine one, but the Console.CursorLeft one.