Screensavers now abort correctly
We added Regen() to manually regenerate the kernel thread. To not break the user code, the Stop() overload now takes the value of whether we need to regenerate the thread or not. This helps in mitigating the race conditions when the said thread requires manual regeneration, such as the case of kernel screensavers.

To not break the user code, we decided that the parameterless Stop() function calls the overloaded Stop() function which we implemented in this commit with the regen variable being set to true.

To start the kernel thread again after manually stopping it with Stop(false), one **must** regenerate the kernel thread using the Regen() function, which basically makes a new instance of our thread before the kernel thread can be used again. We've placed safeguards using the kernel exceptions to prevent any kernel thread user from executing any thread that was manually stopped and not regenerated yet. This throws the KernelException with the error type of KernelExceptionType.ThreadNotReadyYet.

We've also implemented IsReady for the user code to be able to check to see if the said thread is ready to be used or not. It's initially "true" when making a new kernel thread instance.