[RPC-34] [BREAKING] Merging to serialization-based config
A lot of things will have to be done when performing this migration. There's an interesting backstory regarding this change.

When we're developing this version of Nitrocid KS, one of the changes that were to be implemented was that we've had to enhance the configuration reader to stop using the hard-coded values for each config entry. As a result, back on September 9th, 2022, we've implemented the new configuration reader, which was experimental at the time, found in the commit 4fb304a16d435adce12cdf49a2b3f4ce2c4d819a.

The configuration reader was later stabilized to be fit-for-purpose, but we've warned in the commit description that it might be slower than the earliest hard-coded-values-based configuration reader. We've promised that we would be attempting to speed this up across milestones.

The new writer, which was a reflection-based sister of the reflection-based configuration reader, first surfaced on September 20th, 2022 at commit 3348b9d9225f966567565ef66aba66d47bf2978c, and finalized it a day later at commit 55324137381f9039611fb9cdae98b24ba645790a.

We've merged the two configuration reader/writer methods to one, effectively removing the oldest one based on hard-coding the values at commit 0cf5cb1533889df2984c7145731c6e926085098a released on December 23, 2022.

However, the config reader and the writer were slow because it used reflection, so we tried to use the Expressions method to attempt to speed this up, but we've only seen subtle performance improvements, as of commit 73a7c8d3673b01ec26b43f18d17dcb4cb3e411fb released at December 20th, 2022.

This commit will remove the reflection-based config reader and writer and replace them with the serialization-based config reader and writer, and we've noticed significant performance improvements regarding reading the configuration and writing them to a file.

However, as a consequence, the format is changed to satisfy the serialization conditions so that the users can easily pinpoint which part of the config is which. On the main configuration JSON file, we've removed sections to keep the serialization simple.

As a result, we've split the configuration JSON file to three files:

- KernelMainConfig.json
- KernelSaverConfig.json
- KernelSplashConfig.json

However, this implementation is rather incomplete, as the Settings app doesn't understand the changes that the new format required us to do. We've been required to make some fields to be properties, to make some internal changes to some structures, and to fix all the value errors as in the case of the "MaximumColorLevel" fields in some of the screensaver settings, which were set to zero. It also corrects it to 255 so that it can be treated as maximum.

As a result, a couple of removals had to be made. These are the functions from the configuration management module that had to be removed as of this commit:

- GetNewConfigObject()
- ReadFailsafeConfig()
- RepairConfig()
- GetConfigCategory()
- SetConfigValue()
- GetConfigValue()
- GetValueFromEntry()

We've also had to change some of the functions' signatures. Full listing is on the GitBook documentation under "Compatibility" section.

- CreateConfig()
- TryCreateConfig()
- ReadConfig()
- TryReadConfig()

Of course, there may be bugs that we haven't caught during the development, but it's mostly stable.

Good luck!