add - brk|doc - Introduced child threads
The concept of child threads is introduced to Nitrocid KS.

---

Parent threads usually have child threads interacting with them in order to perform operations.

In Nitrocid KS, child threads allow sequential interaction with the parent thread to perform threaded operations in parallel. For example, if Thread1 is a parent thread, Thread2 is another parent thread, Thread3 and Thread4 are child threads of Thread1, and Thread5 and Thread6 are child threads of Thread2, once Thread1 is started, Thread3 and Thread4 are started together. If Thread1 is prompted to stop, Thread3 and Thread4 are stopped, too.

To clarify this, here's a diagram that shows the example of showcasing this concept:

Thread1
|-> Thread3
|-> Thread4
Thread2
|-> Thread5
|-> Thread6

For the Stop() function, however, we added a function that waits for both the parent thread and the child threads to halt their operations for 60000 milliseconds, which is 1 minute. This, along with debugging enabled, tells you which thread failed to stop properly.

For the Wait() function, we've introduced a breaking change that if the base thread is no longer alive (or running), it returns "true". Previously, it returned "false".

---

Type: add
Breaking: True
Doc Required: True
Part: 1/1