fix - reg - Fixed RPC message parsing
Fixes the way that the RPC client parses the messages

---

The RPC client was found to be parsing the command incorrectly, resulting in the client no longer being able to respond to such messages generated by the server. This commit fixes this and fixes the deadlock on shutdown waiting for the RPC thread to shut down.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1