fix - reg - Fixed ProgressClock not setting color until next cycle
ProgressClock had a regression that allowed it to write text in black color.

---

Recent versions of Nitrocid KS have introduced a regression that allowed a screensaver, ProgressClock, to write its information in black color, causing some of the terminals to not to render any text written. This affected the usability of this screensaver, as it wasn't able to display any info until the next color cycle.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1