fix - Fixed crash on file premature end on UESH script parsing
Fixes crash when the file prematurely ended on UESH script parsing

---

When the UESH script parser tries to skip the if condition block because the condition specified in the script didn't satisfy, and encounters the end of file without an empty line in the last line, the script parser crashes. This commit enforces checking for script file line before performing any skips and bails if it encounters end of file.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1