imp - brk|doc - Made Nitrocid KS Extensification-less
We've updated libraries to remove Extensification

---

Our libraries used to include Extensification as a dependency. Since this library is deprecated starting today, we've removed all references to Extensification by updating all libraries and making necessary changes to the API to restore the used Extensification functions to TextTools, IntegerTools, and StreamRead. Custom filesystem drivers in your mods will now have to implement ReadToEndAndSeek() to be compatible with this revision of N-KS, so we've increased the mod API version to 3.0.25.7.

---

Type: imp
Breaking: True
Doc Required: True
Part: 1/1