imp - Improved tmux status height detection logic
Improved tmux status height detection logic by not making assumptions on normal TMUX installs

---

Instead of assuming that the status bar height is 1, we need to actually fetch the "status" variable from the global settings storage in the TMUX program by running a process that gets the value of the said variable so that we can get how many lines of the status bar we have. This way, it'll be able to detect how many lines of the status bar. Applies to TMUX version 2.9 or higher.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1