fix - reg - Fixed TextTools improperly releasing double quotes
TextTools had a regression regarding the usage of the regular expression used to split the command strings with spaces and double quotes.

---

Inconsistency between the two regular expressions aimed to achieving the goal of splitting a string containing spaces and double quotes together created problems, such as the inability for SplitEncloseDoubleQuotes() in the TextTools module to be able to split the string properly. We've made several fixes so that the shell is now able to parse commands with the double quotation marks in the middle of the command properly

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1