imp - prt - Adapted network modules for connections
We've adapted network modules for connections

---

This commit is incomplete, hence we've adapted network modules for network connection tools. We still need to make some refactors in order for the kernel to be able to "detach" the shell from the connection and "attach" it back.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/2