imp - FTP port parsing improved
We've improved logic behind FTP port parsing

---

We used to let the port a string into which we convert later to its integer representation. However, this approach turns out to be ineffective, so we decided to parse this port early into its integer representation so that we don't have to make any conversion later.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1