fix - Fixed crash on PgDn in certain situations
We've fixed a crash in SelectionStyle regarding PgDn and a full page

---

When you have a page that is the last page, and you try to press PgDn to go to the last entry, the SelectionStyle will crash. This fix will check the value and set it accordingly to the last entry's number so that crashes will not happen again.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1