imp - Use build flags to indicate package build
We've used build flags to indicate the version of Nitrocid KS whether it's a package manager version or the compiled version

---

When we're compiling Nitrocid KS for use with package managers, we now check to see if we're building for package managers or standalone. Standalone versions can be distributed with auto update, while the package manager versions can only tell you to update Nitrocid using the package manager.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1