imp - Please, don't set console title twice
We don't need to call Console.Title setter when we already are calling a VT sequence!

---

When we are already calling a VT sequence to change a title, don't use the Console.Title property and then go ahead and call the VT sequence. We don't need to change the title twice.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1