und - Milestones are now alphas
Milestones are now alpha versions

---

Starting with the development of version 0.1.1, which will hopefully start in an undefined date, milestones are now alpha versions. This means that we'll get rid of the Milestone word because it represented fast development, while Alpha is a more comprehensible term.

---

Type: und
Breaking: False
Doc Required: False
Part: 1/1