ref - brk|doc - Renamed Shell to ShellManager
We've renamed Shell to ShellManager for the below reason.

---

The reason as to why we've committed this change is to avoid having to write Shell twice (one for the namespace and one for the class name). This will reduce the total steps taken to reference the class from two to one, effectively allowing you to directly refer to this class, assuming that you've imported the KS.Shell namespace at the top of your mod file.

---

Type: ref
Breaking: True
Doc Required: True
Part: 1/1