imp - prf - Progress and selection no longer flicker
Progress bar writers (horizontal and vertical) and selection style's slider no longer flicker

---

We've improved the progress bar writer performance so that the selection style can take advantage of the gained performance. We now only need to wipe out the relevant areas instead of having the app responsible for rendering the progress bar clear the area. We are the ones who need to clear the area.

This is going to be ported to Terminaux soon.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1