fix - Fixed sproadic hang on modern logon
We've fixed a sproadic hang on modern logon when trying to press ENTER before it shows

---

This can be easily reproduced if you have the RSS feed headline enabled. Previously, the handler would not wait for the logon screen to be fully rendered. Now, we've added a security flag so that once you've tried to press ENTER before the logon screen shows, you'll no longer have to restart the application by clicking on the close button or by killing the dotnet process ("pkill dotnet").

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1