fix - reg - Fixed bizarre rendering crashes
We've fixed some of the crashes caused by the implicit operator

---

The implicit operator defined for Color is causing crashes on some of
the Nitrocid's applications.

Also, we've fixed wrapped writer where it outputs an extra line from the
buffer when it shouldn't. This is to be backported to Terminaux.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1