fix - Fixed StartShellForced() not killing manually exit...
...ing shells

We've fixed StartShellForced() not killing shells that exited manually

---

To verify that this bug is fixed, pass a path to a text file in the "edit -json" command, and it'll indicate that it's exiting the shell. Reverting back to the last shell should be done by forcing KillShellForced() upon newShellCount > shellCount. We made sure that it happens not only when shells throw exceptions, but also when shells exit themselves using the Bail property, such as in the case of the JSON shell.

When we fix this bug, we avoid weird bugs across shells related to shell stacks, like the nasty NullReferenceException bugs or the shells showing weird information or emitting weird messages, depending on the shell either built-in or custom.

Mitigation was done at commit 83ba40f5dcd6a48c4c211114738744a5c3b2cd59 to work around a JSON shell bug causing NRE loop caused by this bug that we've fixed at this commit.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1