fix - brk|doc - Fixed usability bugs related to aliases
We've fixed usability bugs related to aliases.

---

Aliases were not supported by the argument parser for a long time. We've added support for it, as well as fixed a StackOverflowException when trying to use the ShellType enumeration to add aliases, and added DoesAliasExistLocal().

---

Type: fix
Breaking: True
Doc Required: True
Part: 1/1