imp - Handle rooted paths in GetFilesystemEntries()
We've handled rooted paths in GetFilesystemEntries()

---

When GetFilesystemEntries() tries to handle paths that are root directories, we now have an extra check to return paths under the root directory. On Windows, it would be the drive letter followed by the backslash, like C:\, and on Linux, it would be just the slash, which includes folders like boot, bin, etc, usr, var, lib, and so on.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1