imp - Use the MovementDetected property on screen lock
We've added the MovementDetected property to check to see if the cursor has moved.

---

MovementDetected is a property that turns to true if there is a movement. All console drivers that override the following functions must set the _moved property inside the base screensaver class:

  - CursorLeft
  - CursorTop
  - ReadKey()
  - SetCursorPosition()
  - Write()
  - WriteLine()

The following functions must also set this property to true if they don't call the above functions. Most of the time, they don't really need to be overridden because of the improvements that we've done:

  - WritePlain()
  - WriteSlowlyPlain()
  - WriteWherePlain()
  - WriteWhereSlowlyPlain()
  - WriteWrappedPlain()

After this change is done, we've noticed that there are no missed keys in different applications, like apps that use the selection style, interactive TUI, etc.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1